feat: add swing back to checkout
