+++ client/src/components/index.js
@@ -1,4 +1,4 @@
 export { default as Footer } from './Footer'
 export { default as Header } from './Header'
 export { default as LandingFooter } from './LandingFooter'
-export { default as SwingModal } from './SwingModal'
+export { default as SwingModal2 } from './SwingModal2'
